splitted views and serializers
