+++ ngen/views/__init__.py
@@ -1,0 +1,2 @@
+from .auth import *
+from .tools import *
